New List Resource:
# Conflicts:
#	internal/services/network/registration.go
