+++ internal/services/network/registration.go
@@ -200,6 +200,7 @@ func (r Registration) Actions() []func() action.Action {
 func (r Registration) ListResources() []sdk.FrameworkListWrappedResource {
 	return []sdk.FrameworkListWrappedResource{
 		NetworkInterfaceListResource{},
+		NetworkSecurityGroupListResource{},
 		VirtualNetworkListResource{},
 	}
 }
